+++ mrv2/docs/HISTORY.md
@@ -16,7 +16,8 @@ v0.8.2
 - Synchronized menu items to python changes as some were not being taken into
   account.
 - Added imageOptions to session loading and saving.
-- Fixed swallowing of last character in PythonEditor.
+- Fixed swallowing of last character in Python Editor.
+- Fixed nested parenthesis in last expression in Python Editor.
 
 v0.8.1
 ======